# manager_api_getaway/api/docs/swagger.yaml
definitions:
...
  manager_service.CreateManagerTeacher:
    properties:
      branch:
        type: string
      fullname:
        type: string
      group_id:
        type: string
      ielts_attempts_count:
        type: string
      ielts_score:
        type: number
      password:
        type: string
      phone:
        type: string
      salary:
        type: number
    type: object
...
  manager_service.UpdateManagerTeacher:
    properties:
      branch:
        type: string
      fullname:
        type: string
      id:
        type: string
      ielts_attempts_count:
        type: string
      ielts_score:
        type: number
      password:
        type: string
      phone:
        type: string
      role_id:
        type: integer
      salary:
        type: number
    type: object
  models.ResponseError:
    properties:
      error: {}
    type: object
  models.ResponseSuccess:
    properties:
      data: {}
      metadata: {}
    type: object
...
info:
  contact: {}
paths:
...
  /teachers:
    get:
      consumes:
      - application/json
      description: Retrieves information about all teachers
      parameters:
      - description: Search
        in: query
        name: search
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get all teachers
      tags:
      - teachers
    patch:
      consumes:
      - application/json
      description: Update an existing teacher
      parameters:
      - description: Teacher
        in: body
        name: teacher
        required: true
        schema:
          $ref: '#/definitions/manager_service.UpdateManagerTeacher'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Update a teacher
      tags:
      - teachers
    post:
      consumes:
      - application/json
      description: Create a new teacher
      parameters:
      - description: Teacher
        in: body
        name: teacher
        required: true
        schema:
          $ref: '#/definitions/manager_service.CreateManagerTeacher'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create a teacher
      tags:
      - teachers
  /teachers/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a teacher by its ID
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Delete a teacher by its ID
      tags:
      - teachers
    get:
      consumes:
      - application/json
      description: Retrieves a teacher by its ID
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get a teacher by its ID
      tags:
      - teachers
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"